Fixed compilation warning
git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@189813 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acsstartup/src/acsstartupIrFeed.cpp
@@ -127,7 +127,7 @@ void addIdlFiles(std::string dirName,
 int checkIncludeGuards(string idl, string idlPath, multimap<string, string> & includeGuards)
 {
 	string line, temp;
-	int offset;
+	size_t offset;
 	if(idl == "ACSIRSentinel.idl")
 		return 0;
 	ifstream ifs((idlPath+string("/")+idl).c_str());
